reactDjangoV1/requirements.txt
beautifulsoup4
Django
requests
undetected-chromedriver==2.0.0
django-cors-headers
djangorestframework